backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
python-dotenv==1.0.0
supabase==1.0.3
httpx>=0.23.0,<0.24.0
pydantic==1.10.13
python-multipart==0.0.6 